requirements.txt
Django>=4.2
djangorestframework
psycopg2
django-cors-headers
dj-database-url
